django_project/requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cPython==0.0.6
cycler==0.10.0
decorator==4.4.2
Django==2.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
numpy==1.19.1
opencv-contrib-python==4.3.0.36
opencv-python==4.3.0.36
openni==2.3.0
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
primesense==2.2.0.30.post5
progressbar==2.5
prompt-toolkit==3.0.5
protobuf==3.6.0
Pygments==2.6.1
pykalman==0.9.5
pylab-sdk==1.3.2
pymongo==3.11.0
pyparsing==2.4.7
pyrealsense2==2.36.0.2038
python-dateutil==2.8.1
python-osc==1.7.4
pytz==2020.1
pywin32==228
pyzmq==19.0.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow ==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu ==1.14.0
termcolor==1.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
